Refactor Vortex weights loader into modular components
